fastest cp module theme and refactor
