+++ src/EvoSC/EvoSC.csproj
@@ -50,7 +50,7 @@
         <ProjectReference Include="..\Modules\Player\Player.csproj" />
         <ProjectReference Include="..\Modules\PlayerRecords\PlayerRecords.csproj" />
         <ProjectReference Include="..\Modules\SetName\SetName.csproj" />
-        <ProjectReference Include="..\Modules\FastestCp\FastestCp.csproj" />
+        <ProjectReference Include="..\Modules\FastestCpModule\FastestCpModule.csproj" />
         <ProjectReference Include="..\Modules\CurrentMapModule\CurrentMapModule.csproj" />
         <ProjectReference Include="..\Modules\OpenPlanetModule\OpenPlanetModule.csproj" />
         <ProjectReference Include="..\Modules\ModuleManagerModule\ModuleManagerModule.csproj" />
