--- src/Modules/FastestCp/FastestCPModule.cs
@@ -1,8 +1,0 @@
-﻿using EvoSC.Modules.Attributes;
-
-namespace EvoSC.Modules.Official.FastestCp;
-
-[Module(IsInternal = true)]
-public class FastestCpModule : EvoScModule
-{
-}
