--- src/Modules/FastestCp/Models/AccountIdCpTime.cs
@@ -1,3 +1,0 @@
-﻿namespace EvoSC.Modules.Official.FastestCp.Models;
-
-public record AccountIdCpTime(string AccountId, int RaceTime);
