--- src/Modules/FastestCp/Models/PlayerCpTime.cs
@@ -1,3 +1,0 @@
-﻿namespace EvoSC.Modules.Official.FastestCp.Models;
-
-public record PlayerCpTime(string Player, int Cp, TimeSpan Time);
