+++ src/Modules/FastestCpModule/FastestCPModule.cs
@@ -1,0 +1,8 @@
+﻿using EvoSC.Modules.Attributes;
+
+namespace EvoSC.Modules.Official.FastestCpModule;
+
+[Module(IsInternal = true)]
+public class FastestCpModule : EvoScModule
+{
+}
