+++ src/Modules/FastestCpModule/Models/AccountIdCpTime.cs
@@ -1,0 +1,3 @@
+﻿namespace EvoSC.Modules.Official.FastestCpModule.Models;
+
+public record AccountIdCpTime(string AccountId, int RaceTime);
