+++ src/Modules/FastestCpModule/Models/PlayerCpTime.cs
@@ -1,0 +1,3 @@
+﻿namespace EvoSC.Modules.Official.FastestCpModule.Models;
+
+public record PlayerCpTime(string Player, int Cp, TimeSpan Time);
