+++ tests/Modules/FastestCp.Tests/FastestCp.Tests.csproj
@@ -4,7 +4,7 @@
         <TargetFramework>net7.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <RootNamespace>EvoSC.Modules.Official.FastestCp.Tests</RootNamespace>
+        <RootNamespace>EvoSC.Modules.Official.FastestCpModule.Tests</RootNamespace>
         <IsPackable>false</IsPackable>
     </PropertyGroup>
 
@@ -27,7 +27,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\..\src\Modules\FastestCp\FastestCp.csproj" />
+        <ProjectReference Include="..\..\..\src\Modules\FastestCpModule\FastestCpModule.csproj" />
     </ItemGroup>
 
     <ItemGroup>
